Copyable state.
